refactored the config loading code to be re-usable and now the default path for the config is under the user's home
